+++ clang/docs/ReleaseNotes.rst
@@ -125,6 +125,9 @@ Bug Fixes
 - Fixes to builtin template emulation of regular templates.
   `Issue 42102 <https://github.com/llvm/llvm-project/issues/42102>`_
   `Issue 51928 <https://github.com/llvm/llvm-project/issues/51928>`_
+- A SubstTemplateTypeParmType can now represent the pack index for a
+  substitution from an expanded pack.
+  `Issue 56099 <https://github.com/llvm/llvm-project/issues/56099>`_
 
 
 Improvements to Clang's diagnostics